requirements.txt
fastapi
google-cloud-core
google-cloud-storage
gunicorn
httptools
lxml
nest-asyncio
numpy
pandas
python-dotenv
requests
scikit-learn
tmdb
tmdbv3api
urllib3
uvicorn
websockets
uvloop
tensorflow_hub
tensorflow_text
tensorflow
sentence-transformers
faiss-cpu
python-multipart



